add class AbstractDiagnostic
