+++ src/main/java/com/github/_1c_syntax/bsl/languageserver/diagnostics/BSLDiagnostic.java
@@ -28,15 +28,6 @@ import org.eclipse.lsp4j.Diagnostic;
 import java.util.List;
 import java.util.Map;
 
-/**
- * BSLDiagnostic main purpose is to provide collection of LSP {@link Diagnostic},
- * fired on concrete {@link DocumentContext}.
- *
- * Each BSLDiagnostic implementation MUST contain constructor with exactly one parameter {@link DiagnosticInfo}.
- * Passed DiagnosticInfo MUST be stored as a object field and returned by {@link #getInfo()}.
- *
- * {@link #getDiagnostics(DocumentContext)} method SHOULD use {@link DiagnosticStorage} to add and return diagnostics.
- */
 public interface BSLDiagnostic {
 
   List<Diagnostic> getDiagnostics(DocumentContext documentContext);
